Changing the datatype of the int,short from long to original data types.

+++ integration/spark/src/main/scala/org/apache/spark/sql/CarbonDatasourceRelation.scala
@@ -240,8 +240,8 @@ case class CarbonRelation(
         .map(x => AttributeReference(x.getColName, CarbonMetastoreTypes.toDataType(
           metaData.carbonTable.getMeasureByName(factTable, x.getColName).getDataType.toString
               .toLowerCase match {
-            case "int" => "long"
-            case "short" => "long"
+            case "int" => "int"
+            case "short" => "short"
             case "float" => "double"
             case "decimal" => "decimal(" + x.getPrecision + "," + x.getScale + ")"
             case others => others
